+++ Math/IvVector3.cpp
@@ -41,9 +41,9 @@ IvVector3::IvVector3(const IvVector3& other) :
 }   // End of IvVector3::IvVector3()
 
 IvVector3::IvVector3(const double other[]) :
-    x(other[0]),
-    y(other[1]),
-    z(other[2])
+    x((float)other[0]),
+    y((float)other[1]),
+    z((float)other[2])
 {
 }
